+++ changelog.md
@@ -18,6 +18,7 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 ## Fixed
 
 * `smart-contracts` governance actions were failing due too redundant signature when initiated by non-governance wallet
+* Wizards using 'sidechain' in command line parameters are changed to use 'partner-chain' instead
 
 ## Removed